# Closes #13 performance improvements, add more benchmarks, add intersectHash method