Hi!  
I've added more benchmarks and improved performance for some methods.  
Benchmark results without changes were:  
```
goos: darwin
goarch: arm64
pkg: github.com/twharmon/slices
BenchmarkStdLibSortFunc-8   	   98938	     11663 ns/op	    8232 B/op	      11 allocs/op
BenchmarkSortFuncShort-8    	   73940	     13807 ns/op	   16352 B/op	      18 allocs/op
BenchmarkStdLibSort-8       	  118772	      9962 ns/op	    8200 B/op	      10 allocs/op
BenchmarkSortShort-8        	  107100	     10912 ns/op	   16352 B/op	      18 allocs/op
BenchmarkReverse/strings_237_pure-8         	 1987234	       594.0 ns/op	    4096 B/op	       1 allocs/op
BenchmarkReverse/strings_237-8              	    5991	    187634 ns/op	 1238684 B/op	    1878 allocs/op
BenchmarkReverse/strings_7584_pure-8        	   77500	     15337 ns/op	  122880 B/op	       1 allocs/op
BenchmarkReverse/strings_7584-8             	       5	 229619742 ns/op	1780660038 B/op	  113723 allocs/op
BenchmarkReverse/int64_70_pure-8            	14534846	        79.49 ns/op	     576 B/op	       1 allocs/op
BenchmarkReverse/int64_70-8                 	  102580	     11547 ns/op	   55424 B/op	     433 allocs/op
BenchmarkConcat/strings_6x237_pure-8        	  496095	      2189 ns/op	   24576 B/op	       1 allocs/op
BenchmarkConcat/strings_6x237-8             	   47464	     25390 ns/op	  230323 B/op	      59 allocs/op
BenchmarkConcat/int64_6x70_pure-8           	 4193443	       284.0 ns/op	    3456 B/op	       1 allocs/op
BenchmarkConcat/int64_6x70-8                	  317482	      3681 ns/op	   33304 B/op	      49 allocs/op
BenchmarkFilter/strings_237_(len>5)_pure-8  	 1372065	       884.6 ns/op	    4096 B/op	       1 allocs/op
BenchmarkFilter/strings_237_(len>5)-8       	   24979	     47321 ns/op	  261826 B/op	     722 allocs/op
BenchmarkFilter/int64_70_(val>40)_pure-8    	 6286593	       190.3 ns/op	     576 B/op	       1 allocs/op
BenchmarkFilter/int64_70_(val>40)-8         	  219350	      5487 ns/op	   19376 B/op	     218 allocs/op
BenchmarkMap/strings_237_replace_a_b_pure-8 	  223681	      5273 ns/op	    4768 B/op	      99 allocs/op
BenchmarkMap/strings_237_replace_a_b-8      	    5971	    195865 ns/op	 1239359 B/op	    1976 allocs/op
BenchmarkMap/int64_237_*5_pure-8            	 9245161	       128.5 ns/op	     576 B/op	       1 allocs/op
BenchmarkMap/int64_237_*5-8                 	  101328	     11642 ns/op	   55424 B/op	     433 allocs/op
BenchmarkIntersect/int64_70-46-47_pure-8    	  228531	      5262 ns/op	     576 B/op	       1 allocs/op
BenchmarkIntersect/int64_70-46-47-8         	  157063	      7575 ns/op	    6136 B/op	     107 allocs/op
```