@w1kend 2022-07-04
Yes, it does. I'll write some benchmark examples. Basically the original function has complexity O(N<sup>2</sup>). While using hash, complexity would be O(N) but it needs memory allocation for a map. 
Of course speed would be similar on small slices