@w1kend 2022-07-05
Here you can see difference between functions. I've used 3 slices for test.
100 elements in each
```
BenchmarkIntersect/int64_100-100-100_pure-8         	   85341	     13786 ns/op	     896 B/op	       1 allocs/op
BenchmarkIntersect/int64_100-100-100-8              	   86005	     14200 ns/op	     896 B/op	       1 allocs/op
BenchmarkIntersect/int64_100-100-100_hash-8         	   85599	     12987 ns/op	    7540 B/op	       6 allocs/op
```

1000 elements in each
```
BenchmarkIntersect/int64_1000-1000-1000_pure-8         	     939	   1252860 ns/op	    8192 B/op	       1 allocs/op
BenchmarkIntersect/int64_1000-1000-1000-8              	    1006	   1192758 ns/op	    8192 B/op	       1 allocs/op
BenchmarkIntersect/int64_1000-1000-1000_hash-8         	    9535	    121748 ns/op	   66599 B/op	      10 allocs/op
```

10k elements in each
```
BenchmarkIntersect/int64_10000-10000-10000_pure-8         	       8	 127791547 ns/op	   81920 B/op	       1 allocs/op
BenchmarkIntersect/int64_10000-10000-10000-8              	       9	 120289625 ns/op	   81920 B/op	       1 allocs/op
BenchmarkIntersect/int64_10000-10000-10000_hash-8         	    1035	   1153356 ns/op	  917570 B/op	       4 allocs/op
```