@w1kend 2022-07-05
So having several methods you can choose between them. If you don't want to(can't) allocate additional memory then choose first, if slices are quite long then choose second method.