@twharmon 2022-07-07
I think the `IntersectionHash` isn't quite right:

```go
func TestIntersectionHash(t *testing.T) {
	// ...

	t.Run("dupes in slice", func(t *testing.T) {
		a := []string{"foo", "bar", "baz", "foo"}
		b := []string{"baz", "baz"}
		c := []string{"foo", "foo", "foo"}
		want := []string{}
		got := slices.Sort(slices.IntersectionHash(a, b, c))
		assertEqual(t, got, want)
	})
}
```

If you want, we can get your other changes merged in first and add `IntersectionHash` after.